ちゃんと動くようになった

+++ src/main/java/kishida/cnn/activation/ActivationFunction.java
@@ -8,15 +8,15 @@ package kishida.cnn.activation;
 /** 活性化関数 */
 public abstract class ActivationFunction {
 
-    public abstract double apply(double value);
+    public abstract float apply(float value);
 
-    public void applyAfter(double[] values) {
+    public void applyAfter(float[] values) {
         for(int i = 0; i < values.length; ++i){
             values[i] = apply(values[i]);
         }
     }
 
     /** 微分 */
-    public abstract double diff(double value);
+    public abstract float diff(float value);
 
 }
